refactor(studio): 分离 HTTP 生命周期与请求组合
